sdpa: refactor internal primitive into sdpa_test_iface
